+++ tests/visual/helpers.js
@@ -23,12 +23,11 @@ function runDocTest(options, fn) {
         doc.on('end', async () => {
           try {
             const pdfData = Buffer.concat(buffers);
-            const { systemFonts = false } = options;
-            const images = await pdf2png(pdfData, { systemFonts });
+            const images = await pdf2png(pdfData);
             for (let image of images) {
               expect(image).toMatchImageSnapshot({
                 failureThresholdType: 'percent',
-                failureThreshold: 0.01,
+                failureThreshold: options.failureThreshold || 0.01,
               });
             }
             resolve();