+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Bump node version requirement to 20+
 - Bump minimum supported browsers to Firefox 115, iOS/Safari 16
+- Fix text with input x as null
 
 ### [v0.18.0] - 2026-03-14