+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ### Unreleased
 
+- Bump node version requirement to 20+
+- Bump minimum supported browsers to Firefox 115, iOS/Safari 16
+- Fix text with input x as null
+
+### [v0.18.0] - 2026-03-14
+
 - Fix garbled text copying in Chrome/Edge for PDFs with >256 unique characters (#1659)
 - Fix Link accessibility issues
 - Fix Table Accessibility Issue: Operator CS/cs not allowed in this current state